package.json: change optional to mandatory and add signalk-venus-plugin
This changes nothing to the build, as they were already always installed. But
not anymore as per the recent change to npm-online-install.bbclass.

For the reasons why each package is wanted see here:
https://docs.google.com/document/d/1oJLH7MoRAPoqlo9aaO0gnBYA05a2ruQWFsexZfqWm9E

For why we don't want packages to be listed as optional, see here:
https://github.com/victronenergy/venus/wiki/npm-and-nodejs-packages

Part of https://github.com/victronenergy/venus/issues/1304