package.json: no more optional packages, signalk-venus-plugin and more
1) change optional packages to mandatory
This changes nothing to the build, as they were already always installed. But
not anymore as per the recent change to npm-online-install.bbclass.

For the reasons why each package is wanted see here:
https://docs.google.com/document/d/1oJLH7MoRAPoqlo9aaO0gnBYA05a2ruQWFsexZfqWm9E

For why we don't want packages to be listed as optional, see here:
https://github.com/victronenergy/venus/wiki/npm-and-nodejs-packages

Part of https://github.com/victronenergy/venus/issues/1304

2) add signalk-venus-plugin because we need that

3) update @mxtommy/kip and @signalk/freeboard-sk

And do that such that its done automatically when refreshing the shrinkwrap

f